% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingMisCov.R
\name{testingMisCov}
\alias{testingMisCov}
\title{testingMisCov}
\usage{
testingMisCov(trainRes, testData)
}
\arguments{
\item{trainRes}{Results results from trainging results}

\item{testData}{testing data set}
}
\value{
YHat the empirical risk of the test
}
\description{
Compute the empirical risk of cv or testing for kernel machies with missing covariates
}
