% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingMisCovgGeneral.R
\name{trainingMisCovGeneral}
\alias{trainingMisCovGeneral}
\title{trainingMisCovGeneral}
\usage{
trainingMisCovGeneral(
  misCovDat,
  px,
  kerType,
  kerMethod,
  lambda,
  sigma,
  PSFunPath,
  IMPFunPath,
  B,
  testPurpose
)
}
\arguments{
\item{misCovDat}{data set with missing covariates.
(X,V,R,Y1,Y), X is the fully observed covariates; V is the popential missing covariates; R is the missing indicator; Y1=Y the binary response}

\item{px}{dimsension of covariates X (totally observed)}

\item{kerType}{type of kerenl function, "RBF"(if choose) or "linear" (else choose, automatically)}

\item{kerMethod}{method to use "CC", "WCC", and "DR"}

\item{lambda}{tuning parameter}

\item{sigma}{parameter for RBF kernel, also act as tuning parameter}

\item{PSFunPath}{Path of function used to estimated propensity score}

\item{IMPFunPath}{Path of imputation function used to generte imputation data}

\item{B}{The imputation time for DR (B>=1)}

\item{testPurpose}{test purpose, for crossivalidation, the loss function is chosen as the phi loss;
for test (default value) , the loss function is chosen as the classification loss.}
}
\value{
trainRes
}
\description{
training function for kernel machies with non-monotone type of missing covariates
This is train algorithm, we consider three different cases
1. CC; 2.WCC; 3.DR
The algorithm uses the hinge loss function to do minimization procedure
If CC case, return the estimated alpha based on complete data estimated parameters
If WCC case, return the estimated alpha based on complete data and
for propensity score which is also used in testing procedure;
If DR case, return (1) the contrained quadratic program solution alpha
(2) estimated parameters for propensity score (3) estimated parameters
of conditional distribution of x2 given X1 and Y. (4) Covariates of the
dataset after imputation- the large dataset
}
